Prepare release, clean README.md, rename library
