+++ .github/workflows/README.md
@@ -1,0 +1,3 @@
+Check out the guide that was used to create the CI environment including setting up the yaml files:
+
+https://medium.com/@cjolowicz/hypermodern-python-d44485d9d769
