+++ .github/workflows/tests.yml
@@ -10,7 +10,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ['3.6', '3.7', '3.8']
+        python-version: ["3.6", "3.7", "3.8"]
     env:
       OS: ${{ matrix.os }}
       PYTHON: ${{ matrix.python-version }}
