+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# Changelog
+## 0.1.0
+- initial release
+- update README.md
+- update example notebook
+- add Gh Action for release
+- rename library
