+++ tests/download/test_https_handling.py
@@ -1,5 +1,5 @@
 """ Tests for global requests.Session """
-from python_dwd.download.https_handling import create_dwd_session
+from wetterdienst.download.https_handling import create_dwd_session
 
 
 def test_create_dwd_session():
