+++ tests/test_parse_metadata.py
@@ -2,12 +2,12 @@ import pytest
 import requests
 from pandas import Timestamp
 
-from python_dwd import reset_meta_index_cache
-from python_dwd.enumerations.column_names_enumeration import DWDMetaColumns
-from python_dwd.parse_metadata import metadata_for_dwd_data
-from python_dwd.enumerations.parameter_enumeration import Parameter
-from python_dwd.enumerations.period_type_enumeration import PeriodType
-from python_dwd.enumerations.time_resolution_enumeration import TimeResolution
+from wetterdienst import reset_meta_index_cache
+from wetterdienst.enumerations.column_names_enumeration import DWDMetaColumns
+from wetterdienst.parse_metadata import metadata_for_dwd_data
+from wetterdienst.enumerations.parameter_enumeration import Parameter
+from wetterdienst.enumerations.period_type_enumeration import PeriodType
+from wetterdienst.enumerations.time_resolution_enumeration import TimeResolution
 
 
 def test_metadata_for_dwd_data():
