+++ wetterdienst/additionals/util.py
