+++ wetterdienst/download/download_services.py
@@ -3,8 +3,8 @@ from io import BytesIO
 from pathlib import PurePosixPath
 from typing import Union
 
-from python_dwd.download.https_handling import create_dwd_session
-from python_dwd.file_path_handling.path_handling import build_climate_observations_path
+from wetterdienst.download.https_handling import create_dwd_session
+from wetterdienst.file_path_handling.path_handling import build_climate_observations_path
 
 
 def download_file_from_climate_observations(filepath: Union[PurePosixPath, str]) -> BytesIO:
