+++ wetterdienst/download/https_handling.py
@@ -2,7 +2,7 @@ from functools import lru_cache
 import requests
 from requests.adapters import HTTPAdapter
 
-from python_dwd.constants.access_credentials import HTTPS_EXPRESSION
+from wetterdienst.constants.access_credentials import HTTPS_EXPRESSION
 
 MAX_RETRIES = 3
