+++ wetterdienst/enumerations/datetime_format_enumeration.py
