+++ wetterdienst/enumerations/parameter_enumeration.py
