+++ wetterdienst/enumerations/period_type_enumeration.py
