+++ wetterdienst/enumerations/time_resolution_enumeration.py
