+++ wetterdienst/exceptions/failed_download_exception.py
