+++ wetterdienst/run.py
@@ -1,7 +1,7 @@
 """ entrypoints ro tun scripts via Docker or command line """
 import fire
 
-from python_dwd.data_collection import collect_dwd_data
+from wetterdienst.data_collection import collect_dwd_data
 
 
 if __name__ == '__main__':
